pyproject.toml
[tool.poetry]
name = "pair_up"
version = "0.0.1"
description = "Prototype LLM-enhanced Coding Companion"
authors = ["Derek Larson <larson.derek.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.5"
openai = "^0.28"
pandas = "1.5.3"
plotly = "^5.12.0"
python-decouple = "^3.8"
streamlit = "^1.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
